requirements.txt
joblib==1.2.0
matplotlib==3.5.1
numpy==1.22.3
pandas==1.3.5
Pillow==9.4.0
PyYAML==6.0
requests==2.28.1
scikit_learn==1.2.2
seaborn==0.11.2
torch==1.11.0
torchvision==0.12.0
